+++ app/database/managers/user_manager.py
@@ -1,34 +1,48 @@
 import logging
 from uuid import UUID, uuid4
-from app.database.models import Users
+
 # Предполагается, что BaseDBManager в другом файле
 from app.database.managers.abstract_manager import BaseDBManager
+from app.database.models import Users
 
-logger = logging.getLogger('ok_service')
+logger = logging.getLogger("ok_service")
 
 
 class UserManager(BaseDBManager):
-
     @property
     def model(self):
         return Users
 
-    def add_user(self, login, password, name, role, created_by, category=None, city=None):
+    def add_user(
+        self,
+        login,
+        password,
+        name,
+        role,
+        created_by,
+        category=None,
+        city=None,
+        position=None,
+        is_active=True,
+    ):
         password = str(password)  # Принудительная конвертация
         print(f"FINAL PASSWORD BEFORE HASHING: {password}")
         logger.debug(f"Тип пароля при добавлении в бд: {type(password)}")
 
         with self.session_scope() as session:
             city_uuid = UUID(str(city)) if city else None
+            position_uuid = UUID(str(position)) if position else None
             created_by_uuid = UUID(str(created_by))
             new_user = self.model(
                 login=login,
                 name=name,
                 role=role,
-                category=category if category else None,
+                category=category if category is not None else 0,
                 created_by=created_by_uuid,
                 city_id=city_uuid,
-                deleted=False
+                position_id=position_uuid,
+                is_active=is_active,
+                deleted=False,
             )
             new_user.set_password(str(password))  # Здесь хешируется
             # Проверяем хеш
@@ -37,25 +51,28 @@ class UserManager(BaseDBManager):
             try:
                 session.add(new_user)
             except Exception as e:
-                logger.error(f"Ошибка добавления пользователя: {e}", extra={
-                    "login": "database"})
+                logger.error(
+                    f"Ошибка добавления пользователя: {e}", extra={"login": "database"}
+                )
                 raise
 
             return str(new_user.user_id)
 
-    def create_main_admin(self,  password, category=None):
+    def create_main_admin(self, password, category=None, position=None, is_active=True):
         """Добавление нового пользователя с хешированием пароля."""
         with self.session_scope() as session:
             user_id = uuid4()
             new_admin = self.model(
                 user_id=user_id,
-                login='admin',
-                name='admin',
-                role='admin',
-                category=category if category else None,
+                login="admin",
+                name="admin",
+                role="admin",
+                category=category if category is not None else 0,
                 city_id=None,
+                position_id=UUID(str(position)) if position else None,
                 created_by=user_id,
-                deleted=False
+                is_active=is_active,
+                deleted=False,
             )
             # Установим пароль сразу после создания объекта
             new_admin.set_password(password)
@@ -65,16 +82,16 @@ class UserManager(BaseDBManager):
                 # При выходе из контекстного менеджера произойдёт commit
                 return user_id
             except Exception as e:
-                logger.error(f"Ошибка добавления пользователя: {e}",
-                             extra={"login": "database"})
+                logger.error(
+                    f"Ошибка добавления пользователя: {e}", extra={"login": "database"}
+                )
                 raise
 
     def check_password_db(self, username, password):
         """Проверяем пароль пользователя"""
         with self.session_scope() as session:
             try:
-                user = session.query(self.model).filter_by(
-                    login=username).first()
+                user = session.query(self.model).filter_by(login=username).first()
                 if user and user.check_password(password):
                     return True
                 return False
@@ -88,8 +105,11 @@ class UserManager(BaseDBManager):
             user = session.query(self.model).filter_by(user_id=user_id).first()
             if user:
                 user.set_password(new_password)  # Обновляем хэш пароля
-                # Сессия будет закоммичена автоматически при выходе из контекстного менеджера
+                # Сессия будет закоммичена автоматически при
+                # выходе из контекстного менеджера
             else:
-                logger.warning(f"Пользователь с ID {user_id} не найден",
-                               extra={"login": "database"})
+                logger.warning(
+                    f"Пользователь с ID {user_id} не найден",
+                    extra={"login": "database"},
+                )
                 return False